fixed mpu fifo overflow by pooling rx instead of using interrupts

+++ PIDController.h
@@ -6,7 +6,7 @@ class PIDController
 public:
 	PIDController();
 
-	void compute(float state, float goal, float &output);
+	void compute(double state, double goal, double &output);
 	void setTunings(float Kp, float Ki, float Kd);
 	void setTiming(int time);
